get everything to build
