+++ examples/react-client/use-camera-and-microphone/src/DeviceControls.tsx
@@ -1,17 +1,15 @@
 import type { PeerStatus, UserMediaAPI } from "@fishjam-cloud/react-client";
-import type { TrackMetadata } from "./fishjamSetup";
 import type { TrackManager } from "@fishjam-cloud/react-client";
 
 type DeviceControlsProps = {
   status: PeerStatus;
-  metadata: TrackMetadata;
 } & (
   | {
-      device: UserMediaAPI & TrackManager<TrackMetadata>;
+      device: UserMediaAPI & TrackManager<unknown>;
       type: "audio";
     }
   | {
-      device: UserMediaAPI & TrackManager<TrackMetadata>;
+      device: UserMediaAPI & TrackManager<unknown>;
       type: "video";
     }
 );
@@ -20,7 +18,6 @@ export const DeviceControls = ({
   device,
   type,
   status,
-  metadata,
 }: DeviceControlsProps) => {
   const isDeviceStreaming = !!device.getCurrentTrack()?.trackId;
   return (
@@ -69,7 +66,7 @@ export const DeviceControls = ({
         className="btn btn-success btn-sm"
         disabled={status !== "joined" || !device?.stream || isDeviceStreaming}
         onClick={() => {
-          device?.startStreaming(metadata);
+          device?.startStreaming();
         }}
       >
         Stream {type} track
