+++ examples/react-client/videoroom/src/main.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
-import { FishjamContextProvider } from "./client.ts";
+import { FishjamProvider } from "@fishjam-cloud/react-client";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <FishjamContextProvider>
+    <FishjamProvider>
       <App />
-    </FishjamContextProvider>
-  </React.StrictMode>,
+    </FishjamProvider>
+  </React.StrictMode>
 );
