+++ packages/react-client/src/hooks/public.ts
@@ -3,3 +3,4 @@ export { useCamera, useMicrophone } from "./devices";
 export { useParticipants } from "./participants";
 export { useScreenShare } from "./screenShare";
 export { useSelector, useStatus, useTracks, useClient } from "./state";
+export { useSetupMedia } from "./setupMedia";
